Adding package for CPAN distribution Algorithm-Cluster version 1.50 into devel/p5-Algorithm-Cluster.
This module is an interface to the C Clustering Library, a general purpose
library implementing functions for hierarchical clustering (pairwise simple,
complete, average, and centroid linkage), along with k-means and k-medians
clustering, and 2D self-organizing maps. This library was developed at the
Human Genome Center of the University of Tokyo. The C Clustering Library
is distributed along with Cluster 3.0, an enhanced version of the famous
Cluster program originally written by Michael Eisen while at Stanford
University.