lang/gcc-aux: Fix build for PKGSRC_COMPILER=clang
The Link-Time Optimizer requires a couple of functions from the math
library.  For gcc, it appears there are built-in versions that
satisfy the requirement, but building with clang requires an explicit
linkage to libm.

This additional LDFLAG may be required on all platforms when clang is
used for pkgsrc, but until this is confirmed let's limit it to NetBSD.
Fix tested on NetBSD 6.1 amd64 with CLANGBASE=${LOCALBASE}