Add bsdinstall-20130828 to pkgsrc.
This is a portable version of NetBSD's install(1) program, and is intended
to be used as a replacement for the install-sh script on platforms which do
not have a native BSD install program.  The install-sh script does not
support file names with spaces, and this version should be more robust.

This portable version does not include mtree, futimes and chflags support.

Tested in a full SmartOS bulk build with no apparent issues.