Requires cairo-gobject. PKGREVISION -> 2.
This package uses lua-gi at build time to see if it can load
cairo-gobject; this does not (AFAICT) work without buildlinking
lua-gi, and possibly not even then, and I don't want to get into that
unless necessary. So make the failure non-fatal as the package build
system suggests. (Except that the way they suggest isn't working for
some other reason, so do it forcibly.)

I am not really in a position to test it at runtime right now, so if
it doesn't work let me know. At least now the package builds.