Drop nonfunctional cups option, following upstream.
Upstream removed cups support and created a new "cups-filter" package
instead.  Since then, building ghostscript-agpl with the cups option
has failed.  This commit removes the cups option, and should result in
zero changes to packages built without the cups option, and hence does
not PKGREVISION++.

There is still some cups configure code, which is forced off (as it
was without the option).  It's not clear if this configure code should
also be removed (by upstream), or if it's a separate bit of cups
functionality.  Addressing this issue and adding a cups-filter package
is left for future work, after the branch.