Update to 1.3.4:
Xiph.Org is pleased to announce a full release of libvorbis 1.3.4.
The primary feature of this release is reduced static data size in
libvorbisenc, as well as associated minor changes required to
libvorbis and libvorbisfile. The release also contains minor build
fixes and build system updates. There are no functional changes
over the previous 1.3.3 release.